requirements.txt
amqp==5.0.2
appdirs==1.4.3
argon2-cffi==20.1.0
asgiref==3.3.1
autopep8==1.5.4
Babel==2.9.0
billiard==3.6.3.0
CacheControl==0.12.6
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.3.1
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.1.5
django-admin-thumbnails==0.2.5
django-allauth==0.44.0
django-appconf==1.0.4
django-countries==7.0
django-crispy-forms==1.10.0
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-phonenumber-field==5.0.0
django-redis==4.12.1
django-tagulous==1.1.0
gunicorn==20.0.4
html5lib==1.0.1
humanize==3.2.0
idna==2.10
ipaddr==2.2.0
kombu==5.0.2
libsass==0.20.1
lockfile==0.12.2
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
phonenumbers==8.12.15
Pillow==8.1.0
pkg-resources==0.0.0
progress==1.5
prometheus-client==0.8.0
prompt-toolkit==3.0.9
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.0
pyparsing==2.4.6
python-decouple==3.4
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.5
rcssmin==1.0.6
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rjsmin==1.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
